% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{episode}
\alias{episode}
\title{Data on the Handmaid's Tale episodes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46 rows and 8 columns.
}
\source{
\url{"https://en.wikipedia.org/wiki/List_of_The_Handmaid\%27s_Tale_episodes"}
}
\usage{
episode
}
\description{
Data on the Handmaid's Tale episodes
}
\keyword{datasets}
